refact: extracted annotation
