+++ lib/capybara_screenshot_diff/snap_manager.rb
@@ -7,7 +7,6 @@ require "capybara_screenshot_diff/snap"
 
 module CapybaraScreenshotDiff
   class SnapManager
-
     attr_reader :root
 
     def initialize(root)
